Move amd64 interrupt code into its own module
